Move Config to zenoh-router crate
