+++ zenoh-router/src/runtime/mod.rs
@@ -11,12 +11,14 @@
 // Contributors:
 //   ADLINK zenoh team, <zenoh@adlink-labs.tech>
 //
+use std::fmt;
+use std::time::Duration;
 use async_std::sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard};
 use zenoh_util::core::{ZResult, ZError, ZErrorKind};
 use zenoh_util::zerror;
 use zenoh_protocol::core::PeerId;
 use zenoh_protocol::link::Locator;
-use zenoh_protocol::proto::WhatAmI;
+use zenoh_protocol::proto::{WhatAmI, whatami};
 use zenoh_protocol::session::{SessionManager, SessionManagerConfig, SessionManagerOptionalConfig};
 use crate::routing::broker::Broker;
 use crate::runtime::orchestrator::SessionOrchestrator;
@@ -39,8 +41,7 @@ pub struct Runtime {
 
 impl Runtime {
 
-    pub async fn new(version: u8, whatami: WhatAmI, listeners: Vec<Locator>, peers: Vec<Locator>, 
-                    iface: &str, delay: std::time::Duration) -> ZResult<Runtime> {
+    pub async fn new(version: u8, config: Config) -> ZResult<Runtime> {
         let pid = PeerId{id: uuid::Uuid::new_v4().as_bytes().to_vec()};
         log::debug!("Generated PID: {}", pid);
 
@@ -48,7 +49,7 @@ impl Runtime {
 
         let sm_config = SessionManagerConfig {
             version,
-            whatami,
+            whatami: config.whatami,
             id: pid.clone(),
             handler: broker.clone()
         };
@@ -65,8 +66,8 @@ impl Runtime {
         };
 
         let session_manager = SessionManager::new(sm_config, Some(sm_opt_config));
-        let mut orchestrator = SessionOrchestrator::new(session_manager, whatami);
-        match orchestrator.init(listeners, peers, iface, delay).await {
+        let mut orchestrator = SessionOrchestrator::new(session_manager, config.whatami);
+        match orchestrator.init(config).await {
             Ok(()) => {
                 Ok(Runtime { 
                     state: Arc::new(RwLock::new(RuntimeState {
@@ -90,5 +91,79 @@ impl Runtime {
 
     pub async fn close(&self) -> ZResult<()> {
         self.write().await.orchestrator.close().await
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct Config {
+    pub whatami: WhatAmI,
+    pub peers: Vec<Locator>,
+    pub listeners: Vec<Locator>,
+    pub multicast_interface: String,
+    pub scouting_delay: Duration,
+}
+
+impl Config {
+
+    fn default(whatami: WhatAmI) -> Config {
+        Config { 
+            whatami,
+            peers: vec![],
+            listeners: vec![],
+            multicast_interface: "auto".to_string(),
+            scouting_delay: Duration::new(0, 250_000_000),
+        }
+    }
+
+    pub fn new(mode: &str) -> Result<Config, ()> {
+        match mode {
+            "peer" => Ok(Config::peer()),
+            "client" => Ok(Config::client()),
+            _ => Err(()),
+        }
+    }
+
+    pub fn peer() -> Config {
+        Config::default(whatami::PEER)
+    }
+
+    pub fn client() -> Config {
+        Config::default(whatami::CLIENT)
+    }
+
+    pub fn add_peer(mut self, locator: &str) -> Self {
+        self.peers.push(locator.parse().unwrap());
+        self
+    }
+
+    pub fn add_peers(mut self, locators: Vec<&str>) -> Self {
+        self.peers.extend(locators.iter().map(|l| l.parse().unwrap()));
+        self
+    }
+
+    pub fn add_listener(mut self, locator: &str) -> Self {
+        self.listeners.push(locator.parse().unwrap());
+        self
+    }
+
+    pub fn add_listeners(mut self, locators: Vec<&str>) -> Self {
+        self.listeners.extend(locators.iter().map(|l| l.parse().unwrap()));
+        self
+    }
+
+    pub fn multicast_interface(mut self, name: String) -> Self {
+        self.multicast_interface = name;
+        self
+    }
+
+    pub fn scouting_delay(mut self, delay: Duration) -> Self {
+        self.scouting_delay = delay;
+        self
+    }
+}
+
+impl fmt::Display for Config {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        fmt::Debug::fmt(self, f)
     }
 }
